Devices implementation.
* Rename params package to requests package.

* Move JWTClaims function to requests package.

* Remove MAC address field from user model.
